requirements.txt
langchain-astradb
pandas
langchain_google_genai
fastapi 
uvicorn 
python-multipart
jinja2
python-dotenv 
langchain 
langchain_core 
langchain_groq
langchain-google-genai google-generativeai langchain

-e.